TensorFlowMapper transform for scoring Tensorflow models in ML.NET (#704)
* creating dummy file to test permissions. will remove

* test

* TensorFlow scoring, from Zeeshan A.'s branch, with some additional changes.

* creating dummy file to test permissions. will remove

* test

* TensorFlow scoring, from Zeeshan A.'s branch, with some additional changes.

* taking care of review comments; build fixes

* simple change intended to trigger fresh builds to repro OSX-Release build failure

* Prevent input tensors from being GC'ed before TF_SessionRun is called

* Remove Tensorflow models from tests data

Instead bring these from a nuget package created in another repo.

For now this is https://github.com/ericstj/machinelearning-testdata-temp

Soon it will be https://github.com/dotnet/machinelearning-testdata

* Add entry point

* Fix manifest and generated C# API.

* Create a redist package for tensorflow

Create a nuget package that redistributes the TensorFlow C-API.

This is needed because TensorFlow doesn't ship an official NuGet package.

This is a straight up repack of the bits published on tensorflow.org.  I made sure to apply the TensorFlow license to this package and not sign it with our authenticate certificates.

* Make TF redist project a normal MSBuild project

Remove the use of the SDK targets, and define our own build and clean.

* Add TF License file to package

* Don't use fullpaths when un-tar'ing

Tar on windows was failing when msbuild passed it a full path.  Workaround by using relative
paths and running where we extract to.

* Add some logging to TF redist project

* Fix casing of TF redist proj

* Change tests to use redistributed TensorFlow

Also modify TF binding code to use `tensorflow` in its DLLImports.

The runtime will still add the approriate prefix/extension on linux/mac.

* Fix mac / linux tensorflow redist

I was missing the libtensorflow_framework dependency which caused mac and linux tests to fail.

After fixing that, mac still failed due to inability to load libtensorflow_framework.so.

We have to rename these to .dylib to satisfy the CORECLR dllimport convention which broke the internal rpath in libtensorflow which pointed at @rpath/libtensorflow_framework.so.  Fix this by rewriting the renamed libtensorflow.dylib on mac.

Since this operation can only be done on mac, I had to change the build of the redist project to only build the bits appropriate for the building platform.  To make this work correctly in the official build I had to make sure these platform specific builds happen when the native build happens.

* Only include LICENSE if it exists

LICENSE is pulled from the Windows package, so it isn't available on mac or linux.

* unit test to verify TF transform works with ML.NET image transforms

* Factor TensorflowTransform into its own assembly/package

* Update Tensorflow to TensorFlow

* Fix manifest and C# API. Also, add entry point unit test.

* update test case for image transforms; still skipping test for now till we figure out why false positive passes

* Use IDataView instead of var in unit test.

* fix have from Tensorflow to TensorFlow

* Fix unit test to use new TextLoader APIs.

* Add unit test using the pipeline API.

* Validate input dimensions.

* Added XML doc for TensorflowTransform.

* Remove extra dimension for batch size in output.

* Fix input/output validation.

* Introduced BatchSize as constant to replace 1 everywhere.

* enabling unit test of TensorFlowTransform working with ML.NET Image* Transforms

* Extended XML docs with detailed information.

* Ensure we include actual TF license in package

The TF zip/tarballs were missing the actual TF license.  Download this and include it in the package.  Rename the file from the zip/tarballs as THIRD_PARTY_NOTICES.txt as that represents its content.

* Change input validation to validate all dimensions in case of multi dimensional input.

* Corrected typos in the documentation.

* Fix LICENSE inclusion in package.

We didn't define ExtractDirectory on the item and instead had a full path as identity.

This worked on linux/osx since it prepended a ""/ to the path, which was tolerated by the file system (an extra leading slash).

On Windows this doesn't work or course.

Fix by appending to the item that doesn't assume files came from an archive.

* Add symbols package and fix package reference to redist

* Address pull request comments.

* Give more details in input dimension mismatch error message.

* Added a test for LearningPipelineAPI and updated the doc.xml