Replace DropColumns,KeepColumns and ChooseColumns with SelectColumns (#1269)
This adds the SelectColumns Transform and Estimator that is replacing
the DropColumns and ChooseColumns Transforms. With this check-in, Drop
and Choose are still in the code base but will be removed. In order to
support loading older models, SelectColumns supports loading in Drop and
Choose transforms. The changes include:
* Implementation of the SelectColumnsTransform,
SelectColumnsDataTransform and SelectColumnsEstimator
* Backward compatibility with Drop and Choose columns by providing
functions on SelectColumns that will be called when loading the model.
* Entry point apis for calling select from the command line.
* Additional testing of the new functionality.

These changes are related to #754.