Addition of the ValueMappingEstimator and ValueMappingTransform. (#1710)
* Addition of the ValueMappingEstimator and ValueMappingTransform.
This will be replacing the TermLookupTransform and provide a way to
specify the mapping betweeen two values (note this is specified and not
trained). A user can specify the mapping by providing a keys list and
values list that must be equal in size. The Estimator will then generate
a 1-1 mapping based on the two lists.

The PR references #754 which covers the conversion of Transformer to use
the new Estimator API.