Updating MKL (#2867)
Enables OpenMP in both ML.Net and in our usage of Intel's MKL library. This also upgrades the reference of MKL to use version 2018.3.10 for Windows, Linux and Mac. The MKL binaries are packaged in MLNetDeps nuget, with this change we are referencing 0.0.9. 
As a result of these changes, there were other changes that needed to happen:
* Build needed to support openmp, for linux, this updated our docker build containers to have openmp installed as part of the clang compiler installation. For mac, openmp was already installed this now occurs though before the build. 
*  Enabled NumberOfThreads parameter for the SymSGDTrainer, NumberOfThreads now defaults to Environment.ProcessorCount